# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lab Resource Utilization Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: equipmentId
          in: query
          schema:
            type: integer
        - name: userId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{bookingId}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking by ID
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdate"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{bookingId}/approve:
    post:
      operationId: approveBooking
      tags: [bookings]
      summary: Approve a booking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{bookingId}/reject:
    post:
      operationId: rejectBooking
      tags: [bookings]
      summary: Reject a booking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingRejection"
      responses:
        "200":
          description: Rejected booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{bookingId}/cancel:
    post:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
  /bookings/waitlist:
    get:
      operationId: listWaitlist
      tags: [bookings]
      summary: List waitlist entries
      responses:
        "200":
          description: List of waitlist entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WaitlistEntry"
    post:
      operationId: joinWaitlist
      tags: [bookings]
      summary: Join waitlist for equipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WaitlistInput"
      responses:
        "201":
          description: Added to waitlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaitlistEntry"

  # ─── UTILIZATION ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        equipmentId:
          type: integer
        userId:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        status:
          type: string
          enum: [PENDING_APPROVAL, CONFIRMED, IN_USE, COMPLETED, CANCELLED, NO_SHOW]
        purpose:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
        totalCost:
          type: ["number", "null"]
        isRecurring:
          type: boolean
        approvedBy:
          type: ["integer", "null"]
        createdAt:
          type: string
        equipmentName:
          type: ["string", "null"]
        userName:
          type: ["string", "null"]
        departmentName:
          type: ["string", "null"]
      required:
        - id
        - equipmentId
        - userId
        - startTime
        - endTime
        - status
        - isRecurring
        - createdAt
    BookingInput:
      type: object
      properties:
        equipmentId:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        purpose:
          type: string
        notes:
          type: string
        isRecurring:
          type: boolean
      required:
        - equipmentId
        - startTime
        - endTime
    BookingUpdate:
      type: object
      properties:
        startTime:
          type: string
        endTime:
          type: string
        purpose:
          type: string
        notes:
          type: string
    BookingRejection:
      type: object
      properties:
        reason:
          type: string
      required:
        - reason
...
    WaitlistEntry:
      type: object
      properties:
        id:
          type: integer
        equipmentId:
          type: integer
        userId:
          type: integer
        requestedStartTime:
          type: string
        requestedEndTime:
          type: string
        position:
          type: integer
        notified:
          type: boolean
        createdAt:
          type: string
        equipmentName:
          type: ["string", "null"]
        userName:
          type: ["string", "null"]
      required:
        - id
        - equipmentId
        - userId
        - requestedStartTime
        - requestedEndTime
        - position
        - notified
        - createdAt
    WaitlistInput:
      type: object
      properties:
        equipmentId:
          type: integer
        requestedStartTime:
          type: string
        requestedEndTime:
          type: string
      required:
        - equipmentId
        - requestedStartTime
        - requestedEndTime
...